Server/requirements.txt
Flask
flask-cors
numpy
pandas
scikit-learn
matplotlib
jupyter
groq
python-dotenv